parseInt removal

+++ pages/api/getTaxReports.js
@@ -3,6 +3,7 @@
 export default async function handler(req, res) {
 	const { walletAddress } = JSON.parse(req.body);
 	let collatedResult = [];
+
 	for (const walletId of walletAddress) {
 		const txHistory = await fetch(
 			`https://api.etherscan.io/api?module=account&action=txlist&address=${walletId}&startblock=0&endblock=99999999&page=1&offset=50&sort=desc&apikey=T5F14H1RXART3EQSAGZQSU8GJK9P1QRE4C`
@@ -23,9 +24,11 @@ export default async function handler(req, res) {
 		});
 
 		await Promise.all(result).then((values) => (result = values));
-        const results = result.filter((el) => parseInt(el.eth) > 0);
+        const results = result.filter((el) => parseFloat	(el.eth) > 0);
+
 		collatedResult = collatedResult.concat(results);
 	}
+	
 
 	res.status(200).json({ result: [...new Set(collatedResult)] });
 }
